Refactored Logging module for much better readability
